move test to profiling_v2
