+++ tests/profiling/test_profiler.py
@@ -390,27 +390,3 @@ def test_profiler_ddtrace_deprecation():
         from ddtrace.profiling.collector import memalloc  # noqa:F401
         from ddtrace.profiling.collector import stack  # noqa:F401
         from ddtrace.profiling.collector import stack_event  # noqa:F401
-
-
-@mock.patch("ddtrace.internal.telemetry.telemetry_writer.add_log")
-@mock.patch("ddtrace.internal.datadog.profiling.ddup.config")
-@mock.patch("ddtrace.internal.datadog.profiling.ddup.failure_msg", "mock failure message")
-def test_libdd_failure_telemetry_logging(mock_ddup_config, mock_add_log):
-    """Test that libdd initialization failures log to telemetry instead of standard logging"""
-    from ddtrace.internal.telemetry.constants import TELEMETRY_LOG_LEVEL
-
-    test_exception = Exception("Test libdd failure")
-    mock_ddup_config.side_effect = test_exception
-
-    profiler._ProfilerInstance()
-
-    mock_add_log.assert_called_once()
-    call_args = mock_add_log.call_args
-
-    assert call_args[0][0] == TELEMETRY_LOG_LEVEL.ERROR
-
-    message = call_args[0][1]
-    assert "Failed to load libdd" in message
-    assert "falling back to legacy mode" in message
-    assert "Test libdd failure" in message
-    assert "mock failure message" in message
